mv oracle tree traversal to `OracleTable` static
