+++ src/module/dataforged/finding.ts
@@ -1,11 +1,7 @@
-import type { IMove, IOracle, IOracleCategory } from 'dataforged'
+import type { IMove } from 'dataforged'
 import type { IronswornItem } from '../item/item'
 import { cachedDocumentsForPack } from '../features/pack-cache'
-import {
-	SFMoveCategories,
-	SFOracleCategories,
-	ISOracleCategories
-} from './data'
+import { SFMoveCategories } from './data'
 import { hashLookup } from './import'
 
 export async function getFoundryMoveByDfId(
@@ -29,49 +25,4 @@ export async function getDFMoveByDfId(
 		}
 	}
 	return undefined
-}
-
-export function getDFOracleByDfId(
-	dfid: string
-): IOracle | IOracleCategory | undefined {
-	const nodes = findOracleWithIntermediateNodes(dfid)
-	return nodes[nodes.length - 1]
-}
-
-export function findOracleWithIntermediateNodes(
-	dfid: string
-): Array<IOracle | IOracleCategory> {
-	const ret: Array<IOracle | IOracleCategory> = []
-
-	function walkCategory(cat: IOracleCategory): boolean {
-		ret.push(cat)
-
-		if (cat.$id === dfid) return true
-		for (const oracle of cat.Oracles ?? []) {
-			if (walkOracle(oracle)) return true
-		}
-		for (const childCat of cat.Categories ?? []) {
-			if (walkCategory(childCat)) return true
-		}
-
-		ret.pop()
-		return false
-	}
-
-	function walkOracle(oracle: IOracle): boolean {
-		ret.push(oracle)
-
-		if (oracle.$id === dfid) return true
-		for (const childOracle of oracle.Oracles ?? []) {
-			if (walkOracle(childOracle)) return true
-		}
-
-		ret.pop()
-		return false
-	}
-
-	for (const cat of [...SFOracleCategories, ...ISOracleCategories]) {
-		walkCategory(cat)
-	}
-	return ret
 }
